Update robotiqGripper.py
Modified the readAll function. Initially this functions read 6 registers (composed of 2 bytes). In fact there are only 3 registers to read. There was a confusion between registers and byte. Because of this there was some issues with the information retrieved from the registers.
Modification have been done to solve this problem reported on https://dof.robotiq.com/discussion/2471/robotiq-2f-adaptive-gripper-control-with-python but not tested.